Update a few Tau bar scripts
br_08_bar_ambi_bw10_02.thn: Previously this bar script used the correct starsphere but it didn't have the correct white tint that the Bw10 starsphere has. This has now been added for the bar script.
cv_01_bar_ambi_ew01_01.thn: This bar script used a generic nebula as a background. However, this station is not actually in a nebula. The script has been modified so it uses the Ew01 starsphere background.