Update Rügen and Montezuma bar scripts
cv_01_bar_ambi_bw01_03: Fixed the black outlined stars by ensuring the srt_grp of the stars is lower than the starsphere. li_09_bar_ambi_iw04_02: Fixed an issue that caused the asteroids in the background to not render. This was fixed by increasing the usr_flag property of the starsphere