Ensure player takes damage from Mission 7 gas pocket field
For some reason, during Mission 7 when the player reaches the gas pocket field, the player is made invulnerable. Because of this, the player doesn't take any damage from the gas balls, despite Juni saying that they may damage your ship. Hence this could teach the player that the gas balls only show a cool effect and not actually do damage to the ship. This is unfortunately false and so the invulnerable state has been removed. These gas balls specifically do very little damage anyway.