Show hidden infocards for Juneau and Ithaca
In Freelancer's resources there are infocards for many dockable bases, including the ones from Juneau and Ithaca. However, in the released game neither of those bases are dockable. As a result, Freelancer doesn't show the additional info for Juneau and Ithaca. Controls.dll has been edited so it contains merged versions of the Juneau and Ithaca infocard sections.