Fix neuralnet accessory and body for M05 Norfolk announcer
During mission 5 when you talk to the Battleship Norfolk announcer, he has the wrong body. It used to be from Liberty while the base is owned by the Armed Forces. He's now been given a Bretonia body. Additionally, in the second comm he was missing the neuralnet accessory, which he did have in the first comm. This has been added for the second comm too.