Remove helmet from some Juni comm messages in M04
Juni said in two comm messages in Mission 4 that she made it to a few planets in Bretonia. However, while docked she shouldn't be wearing a helmet. Therefore, that's been removed. Furthermore, a delay to the first comm message has been added because you previously couldn't see her in the top left due to being in a jump tunnel.