Rename Tain asteroid field to Skye
The asteroid field next to Perth station in Edinburgh is called 'Tain', while in the respective .ini file, the field is referred to as 'Skye'. Furthermore, some bar rumors also mention 'Skye' rather than 'Tain'. Because I assumed the rumors and ini references are correct, the field has been renamed to 'Skye'.