Add ring to Planet Kitadake
Planet Kitadake's infocard mentions that it has a ring system, while the planet itself doesn't actually have one. This commit adds the Weisser ring to it. In addition to this, the Planet Aso ring type has been changed to Aso, because the Ice ring type doesn't have any cloud effects. The Ice ring was added in JFLP and apart from this, this ring type was never used in vanilla anyway.