Increase Cnd_ProjHit values for Dragon wings in Mission 9
During Mission 9 you often hear the Dragon wings and Ozu saying that the reinforcements are being shot to pieces. Though in the mission script these ships are instructed to self-destruct after being shot only five times, which is a bit strange. This means the reinforcements die long before they actually should be dying. In this commit this number has been increased to 20. In my opinion these lines can be removed altogether but that can always be done later.