Add proper nnvoice anomaly_detected sound
JFLP's patch notes mentioned adding a dummy entry for anomaly_detected to fix a warning. In the FL Beta audio files however, there is a proper audio entry that says 'anomaly detected', which sounds just fine. This has been added to nnvoice.utf. Also the entry for it in voices_recognizable.ini has been given an attenuation value to reduce the loudness and a duration.