Add ku_07 shipdealer THN with corrected player position
The trader position is the same as for equipment dealer in the ku_07_deck THN.
The player for the shipdealer version was however on the wrong spot.
This version puts the player just where it is for the equip dealer to fix it.
This file is not used by any vanilla bases.