Fix rumor about working for Juni
After the first bartender cutscene, there would be a small chance that a rumor would be given about Trent signing on to work with Juni. However, this would be before you even had the chance to talk to her. Now the rumor will only be told after the first mission has been accepted, which chronologically makes the most sense.