Missing Bar Music Duplicates
These files are simply repetitions of the same bar music from previous tracks, but they are present as such in the final game. Currently a fully patched FLHDE install ends up with only the first tracks in their remastered versions, but the others remain their low-bitrate versions from the Disc. This duplicates the HQ versions to overwrite the low bitrate versions present in the music directory.
JUST IN CASE these files are ignored by the entire game anyway, ignore this pull request :)