Fix Freeport 5 bartender telling Corsair rumors instead of Zoner rumors
"We bring Artifacts to this base from Crete. It is not as lucrative a run as others, but the Zoners will find buyers for them. Either way we make our money." -> "We get many of our goods from Freeport 1 in Omega-3. We ship Polymers, Engine Components, and Fertilizers here from that area. The Corsairs must fight the Red Hessians to enter that space, but we may pass unhindered."

"The shipyard Tripoli requires many different Commodities to maintain production of our vessels and raiding crafts. We ship Polymers and Engine Components there from this base." -> "There is a very dangerous radiation zone in the center of the system. Even very good shields will not hold up very long in there. Be careful when traveling through that you avoid it, or you may end up dead."