Fix mispositioned weapon group buttons when HUD is minimized
When the HUD was minimized, the weapon groups would be located at the exact same offset from the shield battery text as when the HUD was not minimized. Since the minimized HUD is less wide, the weapon group buttons were partially outside the minimized HUD borders.
Full patch for the edits (original bytes are from the original non-edited DLL):
01932E: 74 31 DB 2D -> D9 45 E4 B8
019337: 45 E4 89 5D A4 DB 45 A4 D8 4E 24 DE C1 D9 5D E4 9B DB 2D -> 00 D8 48 04 8B 7D FC D9 40 08 D8 9F AC 00 00 00 DF E0 BF
01934E: D8 6D E8 D8 46 20 D9 5D E8 9B -> F6 C4 44 7B 17 D8 67 60 EB 12
0194BC: AE 47 E1 7A 14 AE 47 E1 FC 3F 00 00 7B 14 AE 47 -> A6 9B 44 3D 66 66 A6 3F 0A D7 23 BE A6 9B 44 3C