Undo 4GB virtual memory patch
I'm not convinced that setting the LARGEADDRESSAWARE flag is necessary for Freelancer given that the program always uses little memory. Furthermore, setting this flag for 32-bit programs can supposedly cause slight performance discrepancies. I think it has to do with the chance of two arbitraty objects in memory fitting in cache together is less likely when the addresses are larger. But this is just a guess.