Update m08.ini
In one of my previous commits I made it so that the fourth BD fighter is part of the convoy when you enter Honshu. However, this caused the Naval Forces fighters to spread out around the shipyard, making them easier to kill. This was not my intention as I didn't want to alter the difficulty of the mission in any way. Hence I've decided to revert this, but I also ensured Ozu never says one of the BD fighters has been killed because the rogue BD fighter died right at the start. I really dislike this solution as it doesn't align with the shooting script, but for now I don't know what else to do. Also I adjusted the position of the approaching BD fighter when you enter the Chugoku system as previously it wasn't visible at all, but it should've been.