Replace space, danger and battle music with remastered versions (#224)
Based on Freelancer's official Soundtrack release, those versions were
created as a replacement for the already remade musics. The new versions offer better loops (albeit not perfect) and are less dynamic.
This is to make sure players can hear the different instruments even in
the heat of battle. Where usually a lot of frequencies disappear.
Remastered by Worldwave.