Encrypt dgVoodooCpl exes
The false positives have mostly been resolved with the previously patched dgVoodooCpl_new.exe. However, the dgVoodooCpl_old.exe apparently has a detection rate of 17 according to VirusTotal. I was unable to reduce this amount after patching the exe, so I just decided to encrypt it with a password. Players who actually need to use the exe can just use the provided password to unpack the zip. Moreover, since I don't want to run the risk of antiviruses picking up on dgVoodooCpl_new.exe in the future, I decided to encrypt this file as well.