Fix Fischer being vulnerable during M10
Because Fischer now flies a Humpback, his maneuverability is a lot worse. As a result, it's easier for him to fly into the planets when he's following you. When this happened, Fischer would die and the game would softlock. This commit makes him invulnerable so that he can no longer die.