Change female pilots of Blood Dragons to have Kusari faces and matching body costumes
Blood Dragons seem to have not matching faces in difference to all other Kusari factions. Their heads are all across from the colonies which does not fit their origin.
Also their bodies are not matching their MBase counterparts. Which with the vanilla heads causes their shoulders to be always invisible.