Remake of fix for knowledge sold on Kabukicho (#297)
This particular error in the NPC knowledge offers was first reported in issue #196 and fixed by changing the base to be revealed to be Ohashi Border Station, which is the base mentioned by the dialogue string.

I've reverted the previous fix in favor of updating the dialogue string to the appropriate one instead. Both the Farmers Alliance and Hogosha NPC will properly suggest they will sell you the info on Kagoshima Depot, the Farmers Alliance home base, now.

Reasoning is that NPC knowledge offers of other bases are always done by members of a faction that owns or at least visits the base in question, and neither Hogosha or Farmers Alliance are present on that Police base. The entire intent behind this gameplay feature is likely to give the player ideas of where to go next, often revealing important related trade stations from lawful bars or other hidden pirate bases in pirate bars. Additionally, this fix makes the actual behaviour between HDE and non-HDE players purchasing the rumor identical again. Furthermore, this fix will work for all localisations, just like the previous fix.